SwipeInput class
This MonoBehaviour class attaches to an object with a PlayerInput class, and receives pointer inputs to register quick swipes.
The code will check when the player presses and releases the pointer, and record the position at both those times. Upon releasing, the code will check the distance between the two points (to ensure it’s long enough to not just be a tap), and the time between pressing and releasing (to ensure it’s too short to be a slow drag). If these conditions are met, the time and positions at the start and end of the swipe are recorded in a custom ‘Swipe’ struct, and a UnityEvent is invoked using said struct as a parameter. Events can then be added to the SwipeInput class, in the same manner that one would assign them to a UI element such as a button or slider.
I have tested the code and it seems to do a very good job registering inputs, although I have not ensured that it scales appropriately with different screen sizes.