Static 'Directions' class for dealing with straight directions
Since I'm making a game based around solving a Rubik's cube, a lot of math will be required that revolves around directions along the six main axes in 3D space. As well as a static readonly array of those directions, a static Vector3 function is also included that takes a local directional value and converts it to the closest 'cardinal' direction by comparing dot products. This is included to easily 'round' vague swipes on rotated cubes to the closest axis on the cube.