Work-in-progress RubiksCubeSolver class for player input
This class will allow the player to have input in solving the Rubik’s cube, as well as resetting the game and controlling difficulty by determining the size of the cube. It’s still a work-in-progress, however. The size slider, reset and auto-solve buttons do not work presently, and the swipe controls do not work properly when the cube is viewed from certain angles. These are the top priority problems for me to solve.